Wire Empowered Spell AI tactic and add end-to-end Fireball test
Sorcerer AI now fires Empowered Spell as a bonus-action prime when
sorcery points are available and an enemy sits within spell range
(24 tiles). Slots before Telekinetic in the priority pipeline so
the prime feeds the round's main-action damage spell. Also drop
an unnecessary u32 cast clippy flagged in Magic Missile.

+++ src/actions/spells.rs
@@ -1269,7 +1269,7 @@ impl Action for MagicMissile {
         };
         // 5e upcasting: 3 darts at level 1, +1 dart per level above 1.
         let lvl = crate::engine::action_overrides::cast_level(overrides, 1);
-        let n_darts = (3 + (lvl - 1)) as u32;
+        let n_darts = 3 + (lvl - 1);
         // Roll all dart dice in one call so Empowered Spell metamagic
         // can reroll across the whole spell's pool (RAW: the spell is
         // the unit, not each dart). Non-empowered casters get an
